We now have the functionality of skipping special characters
We simply skip anything between \ and a space when rendering text. This allows us to write text like

Hello, my name is Joh\color=red nn\color=0 y

Then later, on our renderer, we simply set the commands and we can render

'Hello, my name is Johnny' with the n in red.