Started the KBasic project
The idea is to figure out what functions I need.
Create a simple I/O using SDL or just a simple terminal.
Then create a simple BIOS in assembly with the required functions.
And then simply build it all together and run it.

Sounds easy enough. Maybe a week of work?